Switch 'available' return type to uint8_t
This function never returns an error (-1), and the number of bytes available will never be a negative number.

+++ cores/arduino/xinput/USB_XInput_API.cpp
@@ -34,7 +34,7 @@ boolean XInputUSB::connected() {
 	return USBDevice.configured();
 }
 
-int XInputUSB::available() {
+uint8_t XInputUSB::available() {
 	return USB_Available(XINPUT_RX_ENDPOINT);
 }
